Exclude image index from 'pre build image in SBOM'
This commit reduces the scope of the rule
'pre_build_script_task_runner_image_in_sbom' by filtering out image
indexes from the check.
This is because if we ensure that all the references in an image index
are in the SBOM, we don't need to check the image index itself.

Ref: https://issues.redhat.com/browse/EC-1347