Use purl to determine base image registry
A SBOM generator fixed some minor issues with the SBOM document that
affects a base images. This was not compatible with conforma rule. The
new implementation of the rule now depends on the based image purl and
gets the image from there. The spdx type was correctly handled. The
cyclonedx type was fixed in this commit.

JIRA: ISV-5866

Signed-off-by: Ales Raszka <araszka@redhat.com>