Match kebab or snake case in sbom purl
IIUC there was a recent mobster change that introduced the snake
case version of the key. I'm not clear on which one is preferred,
(which may be important depending on other consumers of this data),
but let's assume the newer one is better.

Right now we have do some SBOMs with both, so in the interest of
unblocking people who started getting a
pre_build_script_task_runner_image_in_sbom violation, let's match
both until further notice.

Ref: https://issues.redhat.com/browse/KFLUXSPRT-4463