Make 'inaccessible_parent_*' rules logic clearer
This commit adds the helper function 'labels._has_parent', that checks
if a parent image is specified in the base image annotations.

This helper function is used to make the
'labels.inaccessible_parent_manifest' and
'labels.inaccessible_parent_config' clearer: these rules will only be
evaluated if a parent is actually specified, otherwise they will be
skipped.

This prevents a bug that called builtin oci functions with invalid
parameters (that caused an obscure error, like the one in the linked
support ticket).

Note: we considered adding a rule that checks that the parent image
annotations are always specified in the base image, but it turns out
this is not always the case: bundle image manifests are usually built
'FROM scratch', meaning that they will have empty parent annotations.
Also common image manifests might have the same 'FROM scratch' build.

Ref: https://issues.redhat.com/browse/EC-1402
Ref: https://issues.redhat.com/browse/KFLUXSPRT-4327