Use manifest version annotation for trusted rules
The version extracted from the task ref can't be relied upon, since it
is easily counterfeitable.

Instead, as per ADR decision (see attached link), we are going to
extract the version annotation from inside the task manifest, and use
that to evaluate the version constraints in the trusted task rules.

Ref: https://github.com/konflux-ci/architecture/blob/main/ADR/0054-task-versioning.md#mark-each-meaningful-change-with-a-version-update